# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "COVID Modeling",
    "description": "The COVID Modeling UI was developed to help policy makers explore hypothetical intervention strategies to reduce the impact of COVID-19. It provides a common interface and language to run simulations against multiple models at the same time. This API is intended to facilitate automation, e.g. running weekly simulations with the same parameters, and integration with other tools.",
    "version": "0.0.1",
    "license": {"name": "MIT License", "url": "https://mit-license.org/"}
  },
  "servers": [{"url": "/api"}],
  "security": [{"token": [], "session": []}],
...
  "paths": {
...
    "/simulations/model-runs/{model}": {
      "post": {
        "description": "Schedule new model run",
        "parameters": [
          {
            "in": "path",
            "name": "model",
            "description": "Model slug",
            "schema": {
              "type": "string",
              "$ref": "#/components/schemas/ModelSlug"
            },
            "example": "mrc-ide-covid-sim"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ModelInput"},
              "examples": {
                "custom": {
                  "summary": "Run simulation for a single model using a custom format",
                  "value": {
                    "p": [0.5, 0.25],
                    "u0": [0.99, 0.01, 0],
                    "tspan": [0, 10000]
                  }
                },
                "common": {
                  "summary": "Run simulation for a single model using the common format with complete control over parameters",
                  "value": {
                    "region": "US",
                    "subregion": "US-AK",
                    "parameters": {
                      "calibrationDate": "2020-03-06",
                      "calibrationCaseCount": 0,
                      "calibrationDeathCount": 0,
                      "r0": null,
                      "interventionPeriods": [
                        {
                          "startDate": "2020-03-06",
                          "socialDistancing": "aggressive",
                          "schoolClosure": "aggressive",
                          "caseIsolation": "aggressive",
                          "voluntaryHomeQuarantine": "aggressive",
                          "reductionPopulationContact": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "required": ["id"],
                  "properties": {"id": {"type": "integer"}}
                }
              }
            }
          },
          "422": {
            "description": "Invalid configuration",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {"error": {"type": "string"}}
                }
              }
            }
          }
        },
        "operationId": "postModelRun",
        "tags": ["model-runs"]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Intensity": {
        "type": "string",
        "enum": ["aggressive", "moderate", "mild"],
        "title": "Intensity"
      },
      "InterventionPeriod": {
        "type": "object",
        "properties": {
          "startDate": {
            "$ref": "#/components/schemas/ISODate",
            "description": "An ISO-8601 string encoding the date that these interventions begin."
          },
          "socialDistancing": {
            "$ref": "#/components/schemas/Intensity",
            "description": "The level of social distancing in the region."
          },
          "schoolClosure": {
            "$ref": "#/components/schemas/Intensity",
            "description": "The level of school closure in the region."
          },
          "caseIsolation": {
            "$ref": "#/components/schemas/Intensity",
            "description": "The level to which individuals with symptoms self-isolate."
          },
          "voluntaryHomeQuarantine": {
            "$ref": "#/components/schemas/Intensity",
            "description": "The level to which entire households self-isolate when one member of the household has symptoms."
          },
          "reductionPopulationContact": {
            "type": "number",
            "description": "The estimated reduction in population contact resulting from all of the above interventions.\nSome models require this generalized parameter instead of the individual interventions."
          }
        },
        "required": ["startDate", "reductionPopulationContact"],
        "additionalProperties": false,
        "description": "Each intervention period is specified by a `startDate`, a set of interventions (`socialDistancing`, `caseIsolation`, `voluntaryHomeQuarantine`, and `schoolClosure`), and an estimate of the overall effect of these interventions (`reductionPopulationContact`).\nThis overall estimate is needed because some models do not simulate the effects of individual interventions.\nThe strictness of each intervention is specified roughly, as one of `mild`, `moderate`, or `aggressive`.\nEach model connector is responsible for interpreting this distinction in a way that works for the particular model.",
        "title": "Intervention Period"
      },
      "ISODate": {
        "type": "string",
        "title": "ISO Date",
        "example": "2020-03-06"
      },
...
      "CommonModelInput": {
        "type": "object",
        "properties": {
          "region": {
            "$ref": "#/components/schemas/Region",
            "description": "The desired geographical top-level region for the simulation.\nConnector must check this value (and subregion, if provided) and and either produce results for that geography, or return an error."
          },
          "subregion": {
            "$ref": "#/components/schemas/Subregion",
            "description": "Specific subregion for the simulation.\nMay be omitted"
          },
          "parameters": {
            "$ref": "#/components/schemas/ModelParameters",
            "description": "Desired parameters.\nConnectors may choose to ignore any of these parameters if they wish."
          }
        },
        "required": ["region", "parameters"],
        "additionalProperties": false,
        "description": "A generalized description of the input to an epidemiological model.",
        "title": "Common Model Input"
      },
      "MicroMoBBHRMInput": {
        "type": "object",
        "properties": {
          "p": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Number of patches"
          },
          "tmax": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Number of time steps"
          },
          "aqua_stochastic": {
            "type": "array",
            "items": {"type": "boolean"},
            "description": "Update the aquatic mosquito model stochastically or deterministically?"
          },
          "aqua_molt": {
            "type": "array",
            "items": {"type": "array", "items": {"type": "number"}},
            "description": "Proportion of immature stages which mature and emerge as adults each day\n(rows correspond to patches and columns to days)"
          },
          "aqua_surv": {
            "type": "array",
            "items": {"type": "array", "items": {"type": "number"}},
            "description": "Aquatic daily survival probability\n(rows correspond to patches and columns to days)"
          },
          "aqua_K": {
            "type": "array",
            "items": {"type": "array", "items": {"type": "number"}},
            "description": "Carrying capacity\n(rows correspond to patches and columns to days)"
          },
          "aqua_L": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Initial number of immature mosquitoes (length should equal number of\npatches)"
          },
          "adult_stochastic": {
            "type": "array",
            "items": {"type": "boolean"},
            "description": "Update the adult mosquito model stochastically or deterministically?"
          },
          "adult_f": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Blood feeding rate"
          },
          "adult_q": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Human blood feeding fraction"
          },
          "adult_eip": {
            "type": "array",
            "items": {"type": "number"},
            "description": "the extrinsic incubation period (can be either length 1, 365, or equal to\ntmax)"
          },
          "adult_p": {
            "type": "array",
            "items": {"type": "array", "items": {"type": "number"}},
            "description": "Adult daily survival probability\n(rows correspond to patches and columns to days)"
          },
          "adult_psi": {
            "type": "array",
            "items": {"type": "array", "items": {"type": "number"}},
            "description": "mosquito dispersal matrix (rows must sum to 1)"
          },
          "adult_nu": {
            "type": "array",
            "items": {"type": "number"},
            "description": "number of eggs laid per oviposition"
          },
          "adult_M": {
            "type": "array",
            "items": {"type": "number"},
            "description": "total mosquito density per patch (length should equal number of patches)"
          },
          "adult_Y": {
            "type": "array",
            "items": {"type": "number"},
            "description": "infected mosquito density per patch (length should equal number of patches)"
          },
          "adult_Z": {
            "type": "array",
            "items": {"type": "number"},
            "description": "infectious mosquito density per patch (length should equal number of patches)"
          }
        },
        "required": [
          "p",
          "tmax",
          "aqua_stochastic",
          "aqua_molt",
          "aqua_surv",
          "aqua_K",
          "aqua_L",
          "adult_stochastic",
          "adult_f",
          "adult_q",
          "adult_eip",
          "adult_p",
          "adult_psi",
          "adult_nu",
          "adult_M",
          "adult_Y",
          "adult_Z"
        ],
        "additionalProperties": false,
        "title": "Input for MicroMoB model with Beverton-Holt aquatic mosquito model and\nRoss-Macdonald adult mosquito model."
      },
      "MinimalModelInput": {
        "type": "object",
        "properties": {
          "p": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Vector of model parameters"
          },
          "u0": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Vector of initial conditions"
          },
          "tspan": {
            "type": "array",
            "items": {"type": "number"},
            "description": "Timespan over which to simulate the model"
          },
          "dt": {
            "type": "number",
            "description": "Optional time interval at which to compute the states."
          }
        },
        "required": ["p", "u0", "tspan"],
        "additionalProperties": false,
        "title": "Minimal Model Input"
      },
      "ModelInput": {
        "anyOf": [
          {"$ref": "#/components/schemas/CommonModelInput"},
          {"$ref": "#/components/schemas/MinimalModelInput"},
          {"$ref": "#/components/schemas/MicroMoBBHRMInput"}
        ],
        "title": "Model Input"
      },
...
      "ModelParameters": {
        "type": "object",
        "properties": {
          "calibrationDate": {
            "$ref": "#/components/schemas/ISODate",
            "description": "An ISO-8601 string encoding the date of the most recent case data in the region.\nSome models such as `CovidSim` require you to provide some information about recorded cases in the region, in order to calibrate the model.\nThe web-interface provides this information automatically, based on recorded case data downloaded daily from various international data sources."
          },
          "calibrationCaseCount": {
            "type": "number",
            "description": "The total number of confirmed cases in the region before the calibration date."
          },
          "calibrationDeathCount": {
            "type": "number",
            "description": "The total number of deaths in the region before the calibration date."
          },
          "r0": {
            "type": "number",
            "description": "The assumed [_reproduction number_](https://en.wikipedia.org/wiki/Basic_reproduction_number) for the virus.\nIf this is null, then each model will use its own default value.",
            "nullable": true
          },
          "interventionPeriods": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/InterventionPeriod"},
            "description": "A list of time periods, each with a different set of interventions.\nPolicy interventions are specified as a series of _intervention periods_, each with a certain set of interventions that are in place.\nFor example, case isolation and social distancing may be instituted first, followed by school closure a week later, followed by a relaxation of all guidelines after several months.\n**Note** - In order to specify that _all_ interventions end on a certain date, there should be a _final_ intervention period that starts on that date, has no interventions specified, and has `reductionPopulationContact` set to zero."
          }
        },
        "required": [
          "calibrationDate",
          "calibrationCaseCount",
          "calibrationDeathCount",
          "r0",
          "interventionPeriods"
        ],
        "additionalProperties": false,
        "title": "Model Parameters"
      },
...
      "ModelSlug": {"type": "string"},
...
      "Region": {"type": "string", "title": "Region", "example": "US"},
...
      "Subregion": {"type": "string", "title": "Subregion", "example": "US-AK"}
    },
    "securitySchemes": {
      "token": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Primary authentication method for programmatic use."
      },
      "session": {
        "type": "http",
        "in": "cookie",
        "name": "session",
        "description": "For use in UI. Also requires valid session.sig cookie"
      }
    }
  }
}